Add files via upload
This is the code for humidity tracking system. The humidity_auth.py handles user registration and logins. The humdity_publisher.py simulates a device that publishes humidity data to the MQTT broker every 5 seconds. The humidty_subscriber.py subscribes to the MQTT topic and prints real-time humidity data sent by the publisher.